ruletree serialization
