+++ src/dt.ts
@@ -27,71 +27,7 @@ import {
   toEstree,
 } from "./estree-util";
 
-/*
- * Defines a "Feature" that can be found on a type of Thing, which can
-   have a value of type ValueType. For instance, a Feature<Card,
-   SuitValue>.
- */
-export abstract class Feature<ThingType, ValueType> {
-  value: ValueType;
-  constructor(value: ValueType) {
-    this.value = value;
-  }
-  abstract matches(thing: ThingType): boolean;
-  serialize() {
-    return mkNew(mkIdentifier(this.constructor.name), [toEstree(this.value)]);
-  }
-  equals(other: Feature<any, any>): boolean {
-    if (this.constructor !== other.constructor) {
-      return false;
-    }
-    if (this.value !== other.value) {
-      return false;
-    }
-    return true;
-  }
-}
-
-/*
-  * Defines a Rule that maps a set of Features to a result. The result
-    can be used to store anything.
- */
-export class Rule<T, U> {
-  features: Feature<T, any>[];
-  result: U;
-  constructor(result: U, features: Feature<T, any>[]) {
-    this.result = result;
-    this.features = features;
-  }
-  toString() {
-    return `Rule(features=${this.features})`;
-  }
-}
-
-export class RuleTreeNode<T, U> {
-  feature: Feature<T, any> | null;
-  rules: Rule<T, U>[];
-  left: RuleTreeNode<T, U> | null;
-  right: RuleTreeNode<T, U> | null;
-
-  constructor(
-    feature: Feature<T, any> | null = null,
-    rules: Rule<T, U>[] = [],
-    left: RuleTreeNode<T, U> | null = null,
-    right: RuleTreeNode<T, U> | null = null,
-  ) {
-    this.feature = feature;
-    this.rules = rules;
-    this.left = left;
-    this.right = right;
-  }
-
-  toString() {
-    const feature = this.feature ? this.feature.value : "null";
-    const rules = this.rules.map((r) => r.toString());
-    return `RuleTreeNode(feature=${feature}, rules=[${rules}] left=${this.left}, right=${this.right})`;
-  }
-}
+import { Feature, Rule, RuleTreeNode } from "./definitions";
 
 export function buildRuleTree<T, U>(rules: Rule<T, U>[]): RuleTreeNode<T, U> {
   if (rules.length === 0) {
@@ -153,22 +89,6 @@ export function buildRuleTree<T, U>(rules: Rule<T, U>[]): RuleTreeNode<T, U> {
   }
 
   return node;
-}
-
-function ruleToEstree(rule: Rule<any, any>): Expression {
-  return mkNew(mkIdentifier("Rule"), [
-    toEstree(rule.result),
-    mkArray(rule.features.map((f) => f.serialize())),
-  ]);
-}
-
-export function ruleTreeNodeToEstree(node: RuleTreeNode<any, any>): Expression {
-  return mkNew(mkIdentifier("RuleTreeNode"), [
-    node.feature ? node.feature.serialize() : mkLiteral(null),
-    mkArray(node.rules.map((r) => ruleToEstree(r))),
-    node.left ? ruleTreeNodeToEstree(node.left) : mkLiteral(null),
-    node.right ? ruleTreeNodeToEstree(node.right) : mkLiteral(null),
-  ]);
 }
 
 export function findMatchingRules<T, U>(
